Initial load of sample data from a remote client - only user so far
