+++ .gitignore
@@ -1,0 +1,5 @@
+node_modules
+build
+npm-debug.log
+.env
+.DS_Store
