+++ package.json
@@ -4,7 +4,7 @@
   "description": "Demo of setting up and populating a Cherrytwist Ecoverse",
   "main": "index.js",
   "scripts": {
-    "sample-data": "ts-node-dev  src/test/client-cli.ts"
+    "sample-data": "ts-node-dev  src/submit-sample-data.ts"
   },
   "repository": {
     "type": "git",
@@ -24,6 +24,7 @@
     "@types/node": "^14.6.0",
     "@types/node-fetch": "^2.5.7",
     "graphql": "^15.3.0",
+    "graphql-request": "^3.1.0",
     "node-fetch": "^2.6.0",
     "react": "^16.13.1",
     "typescript": "^4.0.2"
