+++ src/query-variables/create-challenge-grid
@@ -1,0 +1,49 @@
+{
+  "challengeData": 
+{
+    "name": "Balance the grid",
+    "context": {
+        "tagline": "How might we incentivize consumers to communicate energy demand and production to allow all stakeholders to balance the grid?",
+        "background": "Our power system is becoming increasingly more decentralized and complex. By using solar PV, consumers have become electricity producers as well, and the number of these ‘prosumers’ is increasing. All the while, the electricity demand is rising due to the electrification of transport and industry. Cities around the world are electrifying their public transport systems, but if all busses would be charged around the time people are in the kitchen preparing dinner, the current grid would not be able to cope. The energy sector is striving to make the best of these developments, but so far the sector remains fragmented with a large number of initiatives and pilots spread over start-ups, scale-ups, countries, DSOs, TSOs, utilities, and regulators. Balancing the power grid is and will be more and more difficult and expensive. While technological advances have already disrupted many areas of the power system and, for example, empowered consumers to become producers and traders of their own electricity, the balancing markets still resemble an exclusive club of the big industry players. How can we enable also small consumers and prosumers to contribute to balancing? ",
+        "vision": "Enable the entire energy sector to transition from an ego-system to an eco-system, where everyone and anything is incentivized to share demand AND production of green power. You will co-create the fast and energy efficient digital power market commons by developing a protocol communication layer to power an awesome fossil-free future within one generation.",
+        "impact": "The power grid is one of the most sophisticated infrastructures ever built. By building the energy flexibility ecosystem on a digital protocol layer, the network will become fit for a fossil-free future and will prevent major and costly physical grid investments that will impact the grid and its customers. It will make the power system more robust and transparent, enabling anyone to build an infinite amount of business cases on top of your solution.",
+        "who": "Vattenfall is inviting energy specialists throughout the entire sector, from DSOs and TSOs to solar power solution providers, to work with the teams to create the best solutions",
+        "lifecyclePhase": "Defined",
+        "references": [
+            {
+                "name": "video",
+                "uri": "https://youtu.be/-wGlzcjs9CI",
+                "description": "Video explainer for the challenge"
+            },
+            {
+                "name": "visual",
+                "uri": "https://www.odyssey.org/wp-content/uploads/2020/08/1.-Fossil-Fuel-Free-Future-Vattenfall1-72-scaled.jpg",
+                "description": "Visual for the challenge"
+            },
+            {
+                "name": "EnergyWeb2",
+                "uri": "https://www.energyweb.org/",
+                "description": "Official site"
+            }
+        ]
+    },
+    "tags": [
+        {
+            "name": "Java2"
+        },
+        {
+            "name": "Industry2"
+        }
+        ,
+        {
+            "name": "Energy"
+        }
+    ],
+    "groups": [
+        {
+            "name": "firstGroup"
+        }
+    ]
+}
+
+}
