+++ src/query-variables/create-user-ninja
@@ -1,0 +1,15 @@
+{
+  "userData": 
+  {
+    "name": "SuperNinja2",
+    "tags": 
+    [
+      {
+        "name": "Super Awesome"
+      },
+    	{
+      	"name": "Ninja"
+  		}
+    ]
+  }
+}
