+++ src/query-variables/create-userGroup-ninjas
@@ -1,0 +1,21 @@
+{
+  "userGroupData": 
+  {
+    "name": "Teenage Mutant Ninja Turtles",
+    "members": 
+    [
+      {
+        "name" : "Michelangelo"
+      },
+    	{
+      	"name": "Rafaelo"
+      },
+      {
+        "name" : "Leonardo"
+      },
+    	{
+      	"name": "Donatello"
+  		}
+    ]
+  }
+}
